// src/Types.ts
declare module 'highcharts/modules/drilldown';

export interface Category {
    slug: string;
    name: string;
    url: string;
  }
  


export interface Category {
    id: number;
    name: string;
  }
  
  
  export interface Product {
    id: number;
    title: string;
    category: string;
    price: number;
  }
  

  
  